优化的日记卡片的代码结构
